* On error the process exits
* Errors are now displayed in RED
* Info logs are reduced only to the status of the root jar we are validating
